Make UUID tags store Int for Tag compatibility
- Change EntanglementUUID/Update/Delete UUID fields to Int
- generate_pair_uuid now returns Int
- Update test to expect Int
- Fix Tag constructor compatibility with Tag type constraints

+++ test/test_protocolzoo_entanglement_tracker_uuid.jl
@@ -14,7 +14,7 @@
 
     ##
     # Test basic UUID generation
-    @test typeof(generate_pair_uuid()) == UInt128
+    @test typeof(generate_pair_uuid()) == Int
 
     # Test EntanglerProtUUID and basic entanglement creation
